% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completions.R
\name{$.dollar_function}
\alias{$.dollar_function}
\title{Calls a function using the name used after the dollar sign}
\usage{
\method{$}{dollar_function}(x, name)
}
\arguments{
\item{x}{The function to call}

\item{name}{The name to pass in to the function}
}
\value{
The return value of \code{x}
}
\description{
Calls a function using the name used after the dollar sign
}
