% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{check_version}
\alias{check_version}
\title{Asserts that the current EnvironmentModules version is above a specified
version. Raises an error if it is not.}
\usage{
check_version(against = MIN_SUPPORTED_VERSION)
}
\arguments{
\item{against}{The version to check against. By default this is the minimum
version supported by this package.}
}
\value{
Invisible
}
\description{
Asserts that the current EnvironmentModules version is above a specified
version. Raises an error if it is not.
}
\examples{
check_version()
}
