% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completions.R
\name{.DollarNames.module_load}
\alias{.DollarNames.module_load}
\title{Get autocompletions for \code{\link[=module_load]{module_load()}}}
\usage{
\method{.DollarNames}{module_load}(x, pattern)
}
\arguments{
\item{x}{The \code{module_load} function}

\item{pattern}{Ignored}
}
\value{
A character vector of available modules
}
\description{
Get autocompletions for \code{\link[=module_load]{module_load()}}
}
\keyword{autocomplete}
