% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completions.R
\name{.DollarNames.module_unload}
\alias{.DollarNames.module_unload}
\title{Get autocompletions for \code{\link[=module_unload]{module_unload()}}}
\usage{
\method{.DollarNames}{module_unload}(x, pattern)
}
\arguments{
\item{x}{The \code{module_unload} function}

\item{pattern}{Ignored}
}
\value{
A character vector of currently-loaded modules
}
\description{
Get autocompletions for \code{\link[=module_unload]{module_unload()}}
}
\keyword{autocomplete}
