% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low_level.R
\name{get_loaded_modules}
\alias{get_loaded_modules}
\title{Returns a vector of currently loaded environment modules}
\usage{
get_loaded_modules()
}
\value{
A character vector, with one entry per module
}
\description{
Returns a vector of currently loaded environment modules
}
\examples{
get_loaded_modules()
}
\keyword{low_level}
