% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low_level.R
\name{get_module_code}
\alias{get_module_code}
\title{Returns the R code that would execute a given module command}
\usage{
get_module_code(args, env = character())
}
\arguments{
\item{args}{A character vector defining the module command. You do not need
to include the word "module". For example \verb{module load zeromq} in bash
could be converted to \code{run_modulecmd(c("load", "zeromq"))}}

\item{env}{A character vector containing \code{KEY=VALUE} entries
\strong{not as a named vector} defining additional environment variables to set}
}
\value{
An expression object containing the R code produced by this command
}
\description{
Returns the R code that would execute a given module command
}
\examples{
get_module_code("purge") |> eval()
}
