% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low_level.R
\name{get_module_output}
\alias{get_module_output}
\title{Evaluates a module command, and returns the text output.}
\usage{
get_module_output(args, env = character())
}
\arguments{
\item{args}{A character vector defining the module command. You do not need
to include the word "module". For example \verb{module load zeromq} in bash
could be converted to \code{run_modulecmd(c("load", "zeromq"))}}

\item{env}{A character vector containing \code{KEY=VALUE} entries
\strong{not as a named vector} defining additional environment variables to set}
}
\value{
A character vector with class "cli_ansi_string".
}
\description{
Evaluates a module command, and returns the text output.
}
\examples{
get_module_output("list")
}
