% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{get_modulescmd_binary}
\alias{get_modulescmd_binary}
\title{Gets the file path to the \code{modulescmd} executable}
\usage{
get_modulescmd_binary()
}
\value{
A character scalar containing the full file path to the \code{modulescmd}
executable
}
\description{
Gets the file path to the \code{modulescmd} executable
}
