% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low_level.R
\name{get_version}
\alias{get_version}
\title{Get the current Environment Modules version}
\usage{
get_version()
}
\value{
The current Environment Modules version as a \link[base:numeric_version]{base::numeric_version}
object, or NULL if no version can be detected.
}
\description{
This can e.g. be used to enable/disable functionality based
on whether the current version supports his.
}
\examples{
version = check_version()
version < numeric_version("5.0.0")
}
\keyword{low_level}
