% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_level.R
\name{module_avail}
\alias{module_avail}
\title{Lists all modules available to be loaded}
\usage{
module_avail(filter = "")
}
\arguments{
\item{filter}{If provided, a character scalar will be used to filter the
results. Only modules containing \code{filter} as a substring will be returned.}
}
\value{
A character vector whose entries correspond to available to be
loaded
}
\description{
Lists all modules available to be loaded
}
\examples{
module_avail()
}
\keyword{high_level}
