% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{module_load}
\alias{module_load}
\title{Loads one or more environment modules}
\usage{
module_load(
  modules,
  link_libs = FALSE,
  dyn_load_args = list(),
  env = character()
)
}
\arguments{
\item{modules}{A character vector of environment modules to load}

\item{link_libs}{A logical scalar. If TRUE, link R to all the shared
libraries contained within that module. This is not the default behaviour.}

\item{dyn_load_args}{An optional list of arguments to pass to \code{\link[base:dynload]{base::dyn.load()}}
if \code{link_libs} is also \code{TRUE}}

\item{env}{A character vector containing \code{KEY=VALUE} entries
\strong{not as a named vector} defining additional environment variables to set}
}
\value{
A list of \code{DLLInfo} objects, containing the libraries (if any)
that were linked. See \link[base:getLoadedDLLs]{base::getLoadedDLLs} for an explanation of this
class.
}
\description{
Loads one or more environment modules
}
\examples{
module_load("python")
}
