% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_level.R
\name{module_swap}
\alias{module_swap}
\title{Unloads one module, and loads a second module.}
\usage{
module_swap(from, to)
}
\arguments{
\item{from}{A character scalar: the module to unload}

\item{to}{A character scalar: the module to load}
}
\value{
Invisible
}
\description{
Unloads one module, and loads a second module.
}
\details{
Note that this doesn't have all the functionality of \code{\link[=module_load]{module_load()}}
and \code{\link[=module_unload]{module_unload()}}, so you may want to use those for more control.
}
\examples{
module_swap("python/2", "python/3")
}
