% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_level.R
\name{module_unload}
\alias{module_unload}
\title{Unloads one or more environment modules}
\usage{
module_unload(...)
}
\arguments{
\item{...}{Any number of modules to unload as character vectors, which will
all be concatenated together.}
}
\value{
A list of \code{DLLInfo} objects, containing the libraries (if any)
that were linked. See \link[base:getLoadedDLLs]{base::getLoadedDLLs} for an explanation of this
class.
}
\description{
Unloads one or more environment modules
}
\examples{
module_unload("python")
}
