% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low_level.R
\name{run_modulecmd}
\alias{run_modulecmd}
\title{Runs \code{modulecmd} with some arguments. The shell is harcoded as "r" because
(surprise!) that's the language you are using right now.}
\usage{
run_modulecmd(args, ...)
}
\arguments{
\item{args}{A character vector defining the module command. You do not need
to include the word "module". For example \verb{module load zeromq} in bash
could be converted to \code{run_modulecmd(c("load", "zeromq"))}}

\item{...}{Arguments to forward to \code{\link[base:system2]{base::system2()}}}
}
\value{
A character scalar containing the command's output
}
\description{
Runs \code{modulecmd} with some arguments. The shell is harcoded as "r" because
(surprise!) that's the language you are using right now.
}
\details{
This is a low-level unexported function because users are
encouraged to use the higher level functions such as \code{\link[=module_load]{module_load()}}
}
\keyword{low_level}
