% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with.R
\name{with_extra_modules}
\alias{with_extra_modules}
\title{Runs some R code with additional modules loaded (in addition to the
currently loaded modules), without affecting your current modules.}
\usage{
with_extra_modules(new, code)
}
\arguments{
\item{new}{A character vector of modules to load when executing \code{code}}

\item{code}{Any R code to execute}
}
\value{
The result of evaluating \code{code}
}
\description{
Runs some R code with additional modules loaded (in addition to the
currently loaded modules), without affecting your current modules.
}
\examples{
system2("python3", "--version") |> with_extra_modules("python")
}
