% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with.R
\name{with_module_install}
\alias{with_module_install}
\title{Executes a package installation command so that the installed package
will be aware of the modules you currently have loaded.
This means you can load the package without needing to load that module
later, and also that you can load that package now, without restarting your
R session.}
\usage{
with_module_install(expr)
}
\arguments{
\item{expr}{Any R code you want to run that installs packages. Most of the
time this will just be a call to \url{install.packages}}
}
\value{
The result of evaluating \code{expr}
}
\description{
Executes a package installation command so that the installed package
will be aware of the modules you currently have loaded.
This means you can load the package without needing to load that module
later, and also that you can load that package now, without restarting your
R session.
}
\details{
This functionality is performed by setting a temporary value of
the LDFLAGS environment variable in your Makevars file. It will not work
if the R package is not set up to correctly use LDFLAGS, but all
correctly configured R packages will automatically do so.
Also, it is recommended that you have only the minimal number of modules
needed to install this package loaded when you run this.
}
\examples{
options(repos=c(CRAN="https://cloud.r-project.org/"))
module_load("hdf5")
install.packages("hdf5r") |> with_module_install()
}
\keyword{with}
