% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with.R
\name{with_only_modules}
\alias{with_only_modules}
\title{Runs some R code with no modules loaded except for those that you provide,
without affecting your current modules.}
\usage{
with_only_modules(new, code)
}
\arguments{
\item{new}{A character vector of modules to load when executing \code{code}}

\item{code}{Any R code to execute}
}
\value{
The result of evaluating \code{code}
}
\description{
Runs some R code with no modules loaded except for those that you provide,
without affecting your current modules.
}
\examples{
system2("python3", "--version") |> with_only_modules("python", code=_)
}
