% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low_level.R
\name{check_version}
\alias{check_version}
\title{Asserts that the current EnvironmentModules version is above a specified
version. Raises an error if it is not.}
\usage{
check_version(
  against = MIN_SUPPORTED_VERSION,
  action = "for the function you just ran"
)
}
\arguments{
\item{against}{The version to check against. By default this is the minimum
version supported by this package.}

\item{action}{A character scalar describing the action that is being
attempted, in the infinitive conjugation (e.g. "to run X" or "for running Y")}
}
\value{
An invisible value whose value may be changed in the future.
}
\description{
Asserts that the current EnvironmentModules version is above a specified
version. Raises an error if it is not.
}
\examples{
check_version()
}
\keyword{low_level}
