% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{get_module_output}
\alias{get_module_output}
\title{Evaluates a module command, and returns the text output.}
\usage{
get_module_output(args, env = character())
}
\arguments{
\item{args}{A character vector defining the module command. You do not need
to include the word "module". For example \verb{module load zeromq} in bash
could be converted to \code{run_modulecmd(c("load", "zeromq"))}}

\item{env}{A character vector containing \code{KEY=VALUE} entries
\strong{not as a named vector} defining additional environment variables to set}
}
\value{
A character vector with class "cli_ansi_string". If you have the
\href{https://cli.r-lib.org/}{\code{cli}} package installed and loaded, it will enable enhanced
printing of this result. If you don't want to install cli, you should print
this result out using \code{\link[base:cat]{base::cat()}} and \emph{not} \code{\link[base:print]{base::print()}} to ensure the
ANSI control characters are correctly displayed.
}
\description{
Evaluates a module command, and returns the text output.
}
\examples{
get_module_command("list") |> cat()
}
