% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{get_version}
\alias{get_version}
\title{Returns the current Environment Modules version which can be used for
feature testing}
\usage{
get_version()
}
\value{
The current Environment Modules version as a \link[base:numeric_version]{base::numeric_version}
object, or NULL if no version can be detected.
}
\description{
Returns the current Environment Modules version which can be used for
feature testing
}
\examples{
version = check_version()
version < numeric_version("5.0.0")
}
