% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{link_module_libs}
\alias{link_module_libs}
\title{Links R to the shared libraries in a number of directories}
\usage{
link_module_libs(ld_library_paths, ...)
}
\arguments{
\item{ld_library_paths}{A character vector indicating directories to search
in.}

\item{...}{Arguments forwarded to \code{\link[base:dynload]{base::dyn.load()}}}
}
\value{
A list of \code{DLLInfo} objects describing the newly linked shared
libraries
}
\description{
Links R to the shared libraries in a number of directories
}
\examples{
lib_dir = file.path(R.home(), "lib")
link_module_libs(lib_dir)
}
