% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_level.R
\name{module_avail}
\alias{module_avail}
\title{Lists all modules available to be loaded}
\usage{
module_avail(starts_with = NULL, contains = NULL)
}
\value{
A character vector whose entries correspond to available to be
loaded
}
\description{
Lists all modules available to be loaded
}
\examples{
module_avail()
}
\keyword{high_level}
