% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_level.R
\name{module_avail}
\alias{module_avail}
\title{Lists all modules available to be loaded}
\usage{
module_avail(starts_with = NULL, contains = NULL)
}
\arguments{
\item{starts_with}{An optional character scalar. If provided, only modules
whose name starts with this character string will be returned.}

\item{contains}{An optional character scalar. This parameter is only
supported in Environment Modules 4.3+. If provided, it will filter modules
to only those containing this substring.}
}
\value{
A character vector with class "cli_ansi_string".
}
\description{
Lists all modules available to be loaded
}
\examples{
module_avail()
}
