% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{module_list}
\alias{module_list}
\title{Lists all modules that are currently loaded}
\usage{
module_list(starts_with = NULL, contains = NULL, ...)
}
\arguments{
\item{starts_with}{An optional character scalar. If provided, only modules
whose name starts with this character string will be returned.}

\item{contains}{An optional character scalar. This parameter is only
supported in Environment Modules 4.3+. If provided, it will filter modules
to only those containing this substring.}

\item{...}{
  Arguments passed on to \code{\link[=get_module_output]{get_module_output}}
  \describe{
    \item{\code{env}}{A character vector containing \code{KEY=VALUE} entries
\strong{not as a named vector} defining additional environment variables to set}
    \item{\code{args}}{A character vector defining the module command. You do not need
to include the word "module". For example \verb{module load zeromq} in bash
could be converted to \code{run_modulecmd(c("load", "zeromq"))}}
  }}
}
\value{
The same format as \code{\link[=get_module_output]{get_module_output()}}
}
\description{
Lists all modules that are currently loaded
}
\examples{
module_list() |> cat()
}
