% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_level.R
\name{module_list}
\alias{module_list}
\title{Lists all modules that are currently loaded}
\usage{
module_list(filter = "")
}
\arguments{
\item{filter}{If provided, a character scalar will be used to filter the
results. Only modules containing \code{filter} as a substring will be returned.}
}
\value{
A character vector whose entries correspond to modules that are
currently loaded
}
\description{
Lists all modules that are currently loaded
}
\examples{
module_list()
}
\keyword{high_level}
