% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{module_purge}
\alias{module_purge}
\title{Unloads all modules that are currently loaded}
\usage{
module_purge(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=get_module_output]{get_module_output}}
  \describe{
    \item{\code{env}}{A character vector containing \code{KEY=VALUE} entries
\strong{not as a named vector} defining additional environment variables to set}
    \item{\code{args}}{A character vector defining the module command. You do not need
to include the word "module". For example \verb{module load zeromq} in bash
could be converted to \code{run_modulecmd(c("load", "zeromq"))}}
  }}
}
\value{
Invisible
}
\description{
Unloads all modules that are currently loaded
}
\examples{
module_purge()
}
