% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_level.R
\name{module_purge}
\alias{module_purge}
\title{Unloads all modules that are currently loaded}
\usage{
module_purge()
}
\value{
An invisible value whose value may be changed in the future.
}
\description{
Unloads all modules that are currently loaded
}
\examples{
module_purge()
}
