% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_level.R
\name{module_purge}
\alias{module_purge}
\title{Unloads all modules that are currently loaded}
\usage{
module_purge(env = character(0))
}
\arguments{
\item{env}{A character vector containing \code{KEY=VALUE} entries
\strong{not as a named vector} defining additional environment variables to set}
}
\value{
Invisible
}
\description{
Unloads all modules that are currently loaded
}
\examples{
module_purge()
}
