% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_level.R
\name{module_swap}
\alias{module_swap}
\title{Unloads one module, and loads a second module.}
\usage{
module_swap(from, to)
}
\arguments{
\item{from}{A character scalar: the module to unload}

\item{to}{A character scalar: the module to load}
}
\value{
An invisible value whose value may be changed in the future.
}
\description{
Unloads one module, and loads a second module.
}
\examples{
module_swap("python/2", "python/3")
}
\keyword{high_level}
