% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{module_swap}
\alias{module_swap}
\title{Unloads one module, and loads a second module}
\usage{
module_swap(from, to)
}
\value{
Invisible
}
\description{
Unloads one module, and loads a second module
}
\examples{
module_swap("python/2", "python/3")
}
