% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_level.R
\name{module_unload}
\alias{module_unload}
\title{Unloads one or more environment modules}
\usage{
module_unload(...)
}
\arguments{
\item{...}{Any number of modules to unload as character vectors, which will
all be concatenated together.}
}
\value{
An invisible value whose value may be changed in the future.
}
\description{
Unloads one or more environment modules
}
\examples{
module_unload("python")
}
